refactor: simplify rsbuild dev-related configurations
